+++ src/UglyToad.PdfPig/Fonts/CidFonts/Type0CidFont.cs
@@ -21,29 +21,29 @@
         public TransformationMatrix FontMatrix { get; }
         public CidFontType CidFontType => CidFontType.Type0;
         public FontDescriptor Descriptor { get; }
-        public IReadOnlyDictionary<int, decimal> Widths { get; }
+        public IReadOnlyDictionary<int, double> Widths { get; }
 
         public Type0CidFont(ICidFontProgram fontProgram, NameToken type, NameToken subType, NameToken baseFont,
             CharacterIdentifierSystemInfo systemInfo,
-            FontDescriptor descriptor, IReadOnlyDictionary<int, decimal> widths)
+            FontDescriptor descriptor, IReadOnlyDictionary<int, double> widths)
         {
             this.fontProgram = fontProgram;
             Type = type;
             SubType = subType;
             BaseFont = baseFont;
             SystemInfo = systemInfo;
-            var scale = 1 / (decimal)(fontProgram?.GetFontMatrixMultiplier() ?? 1000);
+            var scale = 1d / (fontProgram?.GetFontMatrixMultiplier() ?? 1000);
             FontMatrix = TransformationMatrix.FromValues(scale, 0, 0, scale, 0, 0);
             Descriptor = descriptor;
             Widths = widths;
         }
 
-        public decimal GetWidthFromFont(int characterCode)
+        public double GetWidthFromFont(int characterCode)
         {
             return GetWidthFromDictionary(characterCode);
         }
 
-        public decimal GetWidthFromDictionary(int cid)
+        public double GetWidthFromDictionary(int cid)
         {
             if (cid < 0)
             {
